+++ src/main/java/de/alpharogroup/swing/laf/actions/LookAndFeelAction.java
@@ -32,6 +32,13 @@ import javax.swing.UnsupportedLookAndFeelException;
 
 import de.alpharogroup.swing.laf.LookAndFeels;
 
+/**
+ * The class {@link LookAndFeelAction}.
+ * 
+ * @deprecated use instead the same name class in the new package de.alpharogroup.swing.plaf
+ *             <br>
+ *             Note: will be removed in the next minor release
+ */
 public class LookAndFeelAction extends AbstractAction
 {